[fleche] Support workspace updates
We first need to refactor a bit more document creation error handling.

In this case, we set the document to a `WorkspaceUpdated` state, which
will force a full resumption of the document and the re-creation of
its initial state. Note that this is memoized so low-frequency
re triggers are not a problem.